untested generalisation of fitnesses and counterfactual computations (all double)

+++ include/evolutionary_utils.h
@@ -15,7 +15,7 @@ class Individual {
     Team team;
 public:
     int id;
-    std::vector<int> fitness;
+    std::vector<double> fitness;
     std::vector<double> differenceEvaluations;
     u_int nondominationLevel;
     double crowdingDistance;
@@ -23,7 +23,7 @@ public:
     Individual(const std::string& filename, int id);
     Individual(const std::string& filename, int id, std::vector<Agent> agents);
     // evaluate a team by simulating it and adding the rewards
-    std::vector<int> evaluate(const std::string& filename, std::vector<Environment> environments);
+    std::vector<double> evaluate(const std::string& filename, std::vector<Environment> environments);
     // difference-evaluate the team and update agent-level difference reward
     void differenceEvaluate(const std::string& filename, std::vector<Environment> environments, std::vector<Individual> paretoFront, int paretoIndex, double hypervolume, double lowerBound);
     // return the agents of the team
@@ -46,8 +46,8 @@ public:
     std::vector<Environment> generateTestEnvironments(const std::string& filename);
     Individual binaryTournament(std::vector<std::vector<Individual>> paretoFronts, size_t pSize);
     std::vector<Agent> crossover(Individual parent1, Individual parent2);
-    double getHypervolume(std::vector<Individual> individuals, int hypervolumeOrigin); // computes the hypervolume of the given list of individuals
-    double getHypervolume(std::vector<std::vector<int>> individualFitnesses, int lowerBound); // computes the hypervolume of the given list of individuals
+    double getHypervolume(std::vector<Individual> individuals, double hypervolumeOrigin); // computes the hypervolume of the given list of individuals
+    double getHypervolume(std::vector<std::vector<double>> individualFitnesses, double lowerBound); // computes the hypervolume of the given list of individuals
     bool dominates(Individual a, Individual b); // finds if the individual a dominates individual b
     std::vector<Individual> findParetoFront(const std::vector<Individual>& population); // finds and returns the pareto front in a population
     std::vector<Individual> without(const std::vector<Individual> workingPopulation, const std::vector<Individual> toRemoveSolutions);
